pyproject.toml
[tool.poetry]
name = "travel-router"
version = "0.1.0"
description = ""
authors = ["CSL426 <spark.cs.liao@gmail.com>"]
packages = [
    {include = "feature"}, 
    {include = "main"},
    {include = "database"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
googlemaps = "^4.10.0"
pandas = "^2.2.3"
pydantic = "^2.10.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
openai = "0.28"
qdrant-client = "^1.12.2"
pytest = "^8.3.4"
geopy = "^2.4.1"
line-bot-sdk = "^3.14.2"
flask = "^3.1.0"
pymongo = "^4.10.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
